+++ ShowdownBot/bin/Release/data/roleOverride.js
@@ -1,4 +1,0 @@
-﻿{
-    "suicune":{name:"suicune",role:{setup:true,tank:true},deftype:{physical:true}},
-    "magcargo":{name:"magcargo",item:"airballoon",personal:true,statspread:{spaEV:252,spe:252}}
-}