+++ ShowdownBot/changelog.md
@@ -4,11 +4,13 @@
 ### v0.6.0 / Unreleased ###
 ----
 Planned:
--Weather
--Better Ability support (add "certainAbility", set from ro.js)
 -fix #31
 -fix #39
 
+Added:
+
+* Weather support
+* "certainAbility" attribute for roleOverride pokemon.
 
 ### v0.5.0 / 2016-09-08 ###
 ----